making slow progress with getting search products by userID
