+++ backend/src/index.ts
@@ -19,6 +19,7 @@ import cors from "cors"; // Import the cors middleware
 import upload from "./multerSetup";
 import multer from "multer";
 import productDeleteController from "./controllers/productDeleteController";
+import currentUserProduct from "./controllers/currentUserProduct";
 
 dotenv.config();
 
@@ -55,7 +56,7 @@ app.post(
 app.get("/product", jwtVerification, productListController);
 
 // Define a route for fetching products specific to the current user
-app.get("/currentUser/products", currentUserProductController);
+app.get("/products", currentUserProduct);
 
 
 app.delete("/product/:id", jwtVerification, productDeleteController);
